requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pandas>=2.1.0
tabulate>=0.9.0
colorama>=0.4.6
